requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-filter==1.1.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-filters==0.11.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
Faker==36.1.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
Markdown==3.7
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-gettext==5.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
